Creates tabs for student/teacher registration and adds route in app.js
