+++ package.json
@@ -10,7 +10,9 @@
     "@okta/okta-signin-widget": "^4.1.2",
     "antd": "^4.4.3",
     "axios": "0.21.1",
+    "craco": "0.0.3",
     "craco-less": "^1.17.0",
+    "dotenv": "^8.2.0",
     "json-web-token": "3.2.0",
     "moment": "^2.29.1",
     "moment-timezone": "^0.5.32",
