+++ src/components/common/TeacherRegForm.js
@@ -1,0 +1,345 @@
+import React, { useState } from 'react';
+import { Form, Input, Select, Button } from 'antd';
+import { Menu, Dropdown } from 'antd';
+import { DownOutlined } from '@ant-design/icons';
+import { QuestionCircleOutlined } from '@ant-design/icons';
+import { Tabs } from 'antd';
+import { HomeOutlined, UserOutlined } from '@ant-design/icons';
+import { Typography, Divider } from 'antd';
+
+const options = [];
+const { Title } = Typography;
+const { Option } = Select;
+const { TabPane } = Tabs;
+
+const formItemLayout = {
+  labelCol: {
+    xs: {
+      span: 24,
+    },
+    sm: {
+      span: 8,
+    },
+  },
+  wrapperCol: {
+    xs: {
+      span: 24,
+    },
+    sm: {
+      span: 16,
+    },
+  },
+};
+const tailFormItemLayout = {
+  wrapperCol: {
+    xs: {
+      span: 24,
+      offset: 0,
+    },
+    sm: {
+      span: 16,
+      offset: 8,
+    },
+  },
+};
+
+const layout = {
+  labelCol: {
+    span: 8,
+  },
+  wrapperCol: {
+    span: 16,
+  },
+};
+
+const TeacherRegForm = () => {
+  const [form] = Form.useForm();
+
+  const onFinish = values => {
+    console.log(values);
+  };
+
+  function handleChange(value) {
+    console.log(`selected ${value}`);
+  }
+
+  return (
+    <Tabs defaultActiveKey="2">
+      <TabPane
+        tab={
+          <span>
+            <HomeOutlined />
+            Teacher
+          </span>
+        }
+        key="1"
+      >
+        <Form
+          {...formItemLayout}
+          {...layout}
+          form={form}
+          name="control-hooks"
+          onFinish={onFinish}
+          form={form}
+          name="register"
+          onFinish={onFinish}
+          scrollToFirstError
+        >
+          <h1>Teacher Registration Form</h1>
+
+          <Form.Item
+            name="firstname"
+            label="First Name"
+            rules={[
+              {
+                required: true,
+                message: 'First name required',
+              },
+            ]}
+          >
+            <Input />
+          </Form.Item>
+
+          <Form.Item
+            name="lastname"
+            label="Last Name"
+            rules={[
+              {
+                required: true,
+                message: 'Last name required',
+              },
+            ]}
+          >
+            <Input />
+          </Form.Item>
+
+          <Form.Item
+            name="email"
+            label="Email"
+            rules={[
+              {
+                type: 'email',
+                message: 'The input is not valid E-mail!',
+              },
+              {
+                required: true,
+                message: 'Please input your E-mail!',
+              },
+            ]}
+          >
+            <Input />
+          </Form.Item>
+
+          <Form.Item
+            name="password"
+            label="Password"
+            rules={[
+              {
+                required: true,
+                message: 'Please input your password!',
+              },
+            ]}
+            hasFeedback
+          >
+            <Input.Password />
+          </Form.Item>
+
+          <Form.Item
+            name="confirm"
+            label="Confirm Password"
+            dependencies={['password']}
+            hasFeedback
+            rules={[
+              {
+                required: true,
+                message: 'Please confirm your password!',
+              },
+              ({ getFieldValue }) => ({
+                validator(_, value) {
+                  if (!value || getFieldValue('password') === value) {
+                    return Promise.resolve();
+                  }
+
+                  return Promise.reject(
+                    'The two passwords that you entered do not match!'
+                  );
+                },
+              }),
+            ]}
+          >
+            <Input.Password />
+          </Form.Item>
+
+          <Form.Item name="location" label="Location">
+            <Input />
+          </Form.Item>
+
+          <Form.Item
+            name="phone"
+            label="Phone Number"
+            rules={[
+              {
+                required: true,
+                message: 'Please input your phone number!',
+              },
+            ]}
+          >
+            <Input
+              style={{
+                width: '100%',
+              }}
+            />
+          </Form.Item>
+          <Form.Item {...tailFormItemLayout}>
+            <Button type="primary" htmlType="submit">
+              Register
+            </Button>
+          </Form.Item>
+        </Form>
+      </TabPane>
+      <TabPane
+        tab={
+          <span>
+            <UserOutlined />
+            Student
+          </span>
+        }
+        key="2"
+      >
+        <Form
+          {...formItemLayout}
+          {...layout}
+          form={form}
+          name="control-hooks"
+          onFinish={onFinish}
+          form={form}
+          name="register"
+          onFinish={onFinish}
+          scrollToFirstError
+        >
+          <h1>Student Registration Form</h1>
+          <Form.Item
+            noStyle
+            shouldUpdate={(prevValues, currentValues) =>
+              prevValues.gender !== currentValues.gender
+            }
+          >
+            {({ getFieldValue }) => {
+              return getFieldValue('gender') === 'other' ? (
+                <Form.Item
+                  name="customizeGender"
+                  label="Customize Gender"
+                  rules={[
+                    {
+                      required: true,
+                    },
+                  ]}
+                >
+                  <Input />
+                </Form.Item>
+              ) : null;
+            }}
+          </Form.Item>
+
+          <Form.Item
+            name="name"
+            label="Name"
+            rules={[
+              {
+                required: true,
+                message: 'Please input your first & last name',
+              },
+            ]}
+          >
+            <Input />
+          </Form.Item>
+
+          <Form.Item
+            name="email"
+            label="Email"
+            rules={[
+              {
+                type: 'email',
+                message: 'The input is not valid E-mail!',
+              },
+              {
+                required: true,
+                message: 'Please input your E-mail!',
+              },
+            ]}
+          >
+            <Input />
+          </Form.Item>
+
+          <Form.Item
+            name="password"
+            label="Password"
+            rules={[
+              {
+                required: true,
+                message: 'Please input your password!',
+              },
+            ]}
+            hasFeedback
+          >
+            <Input.Password />
+          </Form.Item>
+
+          <Form.Item
+            name="confirm"
+            label="Confirm Password"
+            dependencies={['password']}
+            hasFeedback
+            rules={[
+              {
+                required: true,
+                message: 'Please confirm your password!',
+              },
+              ({ getFieldValue }) => ({
+                validator(_, value) {
+                  if (!value || getFieldValue('password') === value) {
+                    return Promise.resolve();
+                  }
+
+                  return Promise.reject(
+                    'The two passwords that you entered do not match!'
+                  );
+                },
+              }),
+            ]}
+          >
+            <Input.Password />
+          </Form.Item>
+
+          <Form.Item name="location" label="Location">
+            <Input />
+          </Form.Item>
+
+          <Form.Item
+            name="phone"
+            label="Phone Number"
+            rules={[
+              {
+                required: true,
+                message: 'Please input your phone number!',
+              },
+            ]}
+          >
+            <Input
+              style={{
+                width: '100%',
+              }}
+            />
+          </Form.Item>
+          <Form.Item {...tailFormItemLayout}>
+            <Button type="primary" htmlType="submit">
+              Register
+            </Button>
+          </Form.Item>
+        </Form>
+      </TabPane>
+    </Tabs>
+  );
+};
+
+export default TeacherRegForm;
